requirements.txt
fastapi
Pillow
pyTelegramBotAPI
sdnotify
uvicorn
